Added the Editor to startup along with the engine for now.
Right now the Editor is a simple proof of concept. It has a couple
different controls, and has event listeners detecting clicks on the
window, blue panel, and green button (the message is printed to the
command line for now). I'm working towards having a panel that will
detect where you click, and put an object there, which will then be
serialized with the world.