#27 SpriteSheets images are now split up in the editor
I need to do a little bit of refactoring so that the editor / game
engine call the same image parsing function. The difference is that
afterwards the engine turns the image into an OpenGL texture, whereas
the editor turns it into an icon.

Also - I renamed a function that used incorrect casing.