#27 Added Preview Panel. SpriteSheets can be subdivided now.
Learned that I should be making these UI components as private member
variables so they are accessible outside of the method they are created
in... kind of obvious, but now I need to do some refactoring. Sprites
can be dynamically subdivided now via the UI - however, you currently
need to press "enter" for the event receiver on the textfields to
trigger.

Also, slightly refactored some Image code so that the editor/engine call
the same code, rather than duplicating in the editor.