Changed the functionality of the "S" key to serialize to a file
#28 The world will now serialize to a file with the name world_xx.xml
where xx is an integer. It will try to detect similar files and
increment the xx number so that they are not over written.

Currently the engine still looks at the "world.xml" file, so if we
wanted to make it look at the latest "world_xx.xml" file then that will
need to be changed.